backend/pyproject.toml
[project]
name = "xai-report-builder-backend"
version = "0.1.0"
description = "Backend for XAI Report Builder"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "celery[redis]>=5.4.0",
  "fastapi>=0.115.0",
  "openpyxl>=3.1.5",
  "pgvector>=0.3.6",
  "pydantic-settings>=2.7.0",
  "pypdf>=5.1.0",
  "pyjwt>=2.10.0",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
  "psycopg[binary]>=3.2.3",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.3",
]
local-ai = [
  "sentence-transformers>=3.4.1",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning:fastapi\\.routing",
  "ignore::DeprecationWarning:starlette\\._utils",
]
